+++ src/data.rs
@@ -28,13 +28,20 @@ pub struct DiffData {
     pub stats: Stats,
 }
 
+impl DiffData {
+    pub fn file_tree(&self) -> Tree {
+        let paths = self.diffs.iter().map(|d| d.file_name()).collect();
+
+        Tree::new(paths)
+    }
+}
+
 #[derive(PartialEq, Clone, Default)]
 pub enum DiffType {
     #[default]
     Modified,
     Staged,
 }
-
 impl DiffType {
     pub fn label_text(&self) -> String {
         match self {
@@ -87,4 +94,88 @@ pub enum Message {
     ChangeDiffType(DiffType),
     ChangeSelectedDiffIndex(usize),
     CloseError,
+}
+
+#[derive(Debug)]
+pub struct Tree {
+    pub nodes: Vec<Tree>,
+    pub files: Vec<String>,
+    pub name: String,
+}
+
+impl Tree {
+    fn new(paths: Vec<PathBuf>) -> Self {
+        let mut tree = Tree {
+            nodes: vec![],
+            files: vec![],
+            name: "".to_owned(),
+        };
+
+        for path in paths {
+            tree.add(path, 0);
+        }
+
+        tree
+    }
+
+    fn add(&mut self, path: PathBuf, depth: usize) {
+        // base cases
+
+        // top level
+        if path.components().count() == 1 {
+            self.files.push(path.to_str().unwrap().to_owned());
+            return;
+        }
+
+        // deepest level
+        if path.components().count() == depth {
+            self.files
+                .push(path.file_name().unwrap().to_str().unwrap().to_owned());
+            return;
+        }
+
+        let name = path
+            .components()
+            .nth(depth)
+            .unwrap()
+            .as_os_str()
+            .to_str()
+            .unwrap()
+            .to_owned();
+
+        // do we already have a tree for this?
+        for node in &mut self.nodes {
+            if node.name == name {
+                node.add(path.clone(), depth + 1);
+                return;
+            }
+        }
+
+        // create a new tree
+        let mut tree = Tree {
+            nodes: vec![],
+            files: vec![],
+            name,
+        };
+        tree.add(path, depth + 1);
+        self.nodes.push(tree);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_tree() {
+        let paths = vec![
+            PathBuf::from(r"src/data.rs"),
+            PathBuf::from(r"src/test.rs"),
+            PathBuf::from(r"src/ui/file_area.rs"),
+        ];
+
+        let tree = Tree::new(paths);
+
+        dbg!(tree);
+    }
 }
